Refactor manifest module into separate files
